+++ src/index.tsx
@@ -3,13 +3,13 @@
 
 import * as React from "react";
 import { render } from "react-dom";
-import { ThemeStyle, ITeamsComponentProps, ITeamsComponentState } from "msteams-ui-components-react";
+import { ThemeStyle } from "msteams-ui-components-react";
 import * as microsoftTeams from "@microsoft/teams-js";
 
 /**
  * State interface for the Teams Base user interface React component
  */
-export interface ITeamsBaseComponentState extends ITeamsComponentState {
+export interface ITeamsBaseComponentState {
     /**
      * The font size
      */
@@ -24,7 +24,7 @@ export interface ITeamsBaseComponentState extends ITeamsComponentState {
 /**
  * Properties interface for the Teams Base user interface React component
  */
-export interface ITeamsBaseComponentProps extends ITeamsComponentProps {
+export interface ITeamsBaseComponentProps {
 
 }