+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Changed
+* Updated Typescript, React, teams-js and msteams-ui-components-react libraries
+
 ## [*1.0.0*] - <*2018-08-13*>
 
 ### Changed